feat: rename header to head
