+++ frontend/src/lib/components/head.svelte
