+++ frontend/src/routes/+page.svelte
@@ -4,7 +4,7 @@
   import Select from "svelte-select"
   import MediaSearch from "$lib/components/media_search.svelte"
   import Filters from "$lib/components/filters.svelte"
-  import Header from "$lib/components/header.svelte"
+  import Head from "$lib/components/head.svelte"
   import { currentCountry } from "$lib/stores/country.js"
   import { currentProviders } from "$lib/stores/providers.js"
   import { currentGenres } from "$lib/stores/genres"
@@ -138,7 +138,7 @@
   })
 </script>
 
-<Header
+<Head
   title="streamchaser"
   description="Trying to keep track of where to watch your favorite movies and TV series? Streamchaser provides the
   definitive guide to what each streaming service has on offer right now!"
